Rename header file for clarity.
